backend/requirements.txt
Django==5.2.3
djangorestframework==3.16.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.2.0 